Working on the services to start at boot time.
The services work. The commands in them work.
But I still have troubles after a reboot.

The gpg-socket gets created in the wrong location, because the runtime
directory is updated.

I thought it was done by the systemd-logind.service and the
user@1001.service, and I made the ega services start after, but
apparently, I'm wrong. It does not seem to work that way.