Trying to update the testsuite
U.5 and F.3 are still failing because the inbox was not removed,
when it should.

An inbox is a fuse mountpoint now, so fusermount -u <inbox> or
umount <inbox> is necessary and apparently the device is busy.
This should not be the case, so I wonder if the SFTP disconnect
really worked.